backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paho-mqtt==2.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bleak==0.22.3
